Feature/csi 145 add storage configuration to csi test pipeline (#24)
This is a ticket to add the usage of ansible to create the storage configuration needed for the csi test pipeline.
this included adding 2 new stages to the pipline - environment setup + storage setup.
for the environment setup the env vars that say where is the yaml file for the storage configuration , and this yaml file is updated with the environments username + password .
besides this the pool name is also taken from the configuration yaml, adn used later for the csi tests.
in the storage configuration the yaml file is passed to a groovy script : configureStorage which creates all the storage entities.

after the tests complete there are some volumes left in the storage pool and ansible right now does not have the capability to delete them so I added a new python script that deletes the left volumes in the pool (so the next pipeline run will be successful). this script is run in a rhel 7 container with python.