Feature/csi node docker build in CI pipeline (#27)
Build csi node as docker image and trigger it as part of the build pipeline in jenkins. 
So from now on there is ready csi node image in the internal artifactory after every push in github (together with csi controller image which already there).

The pipeline also has a unit testing phase for csi node (runs as a container).